Extract usePhaseFade and useReelHaptics from SpinReel
The phase-fade/winner-pulse and slot-crossing haptic effects were two
unrelated side effects living next to the imperative API. Move each
into its own hook and add reduced-motion handling that the in-line
versions lacked: phaseFade jumps to its target without animation,
winnerPulse stays at 1, and haptics are suppressed (the same signal
that asks for less motion usually asks for less vibration).

SpinReel is now 107 lines and contains JSX plus three hook calls.
